рефакторинг модуля генерации случайных данных
