--- src/model/event-points-model.js
@@ -1,17 +1,0 @@
-import {
-  getRandomEventPoint,
-} from '../mock/event-points.js';
-
-import {
-  POINTS_COUNT,
-} from '../const.js';
-
-export default class PointsModel {
-  points = Array.from({
-    length: POINTS_COUNT
-  }, getRandomEventPoint);
-
-  getEventPoints() {
-    return this.points;
-  }
-}
