+++ src/model/points-model.js
@@ -1,0 +1,10 @@
+export default class PointsModel {
+  constructor(service) {
+    this.service = service;
+    this.points = this.service.getPoints();
+  }
+
+  getPoints() {
+    return this.points;
+  }
+}
