+++ src/presenter/points-presenter.js
@@ -1,8 +1,8 @@
 import {
   render
 } from '../render.js';
-import EventListView from '../view/events-list-view.js';
-import EventItemView from '../view/event-item-view.js';
+import EventListView from '../view/event-list-view.js';
+import PointView from '../view/point-view.js';
 import PointEditorView from '../view/point-editor-view.js';
 import SortingView from '../view/sorting-view.js';
 
@@ -18,25 +18,22 @@ export default class PointsPresenter {
   }) {
     this.tripEventsContainer = tripEventsContainer;
     this.pointsModel = pointsModel;
-    this.offersModel = offersModel;
     this.destinationsModel = destinationsModel;
+    this.offersModel = offersModel;
   }
 
   init() {
-    this.eventPoints = [...this.pointsModel.getEventPoints()];
-
+    this.points = [...this.pointsModel.getPoints()];
     render(this.sortingComponent, this.tripEventsContainer);
     render(this.listComponent, this.tripEventsContainer);
     render(new PointEditorView(), this.listComponent.getElement());
 
-    for (let i = 0; i < this.eventPoints.length; i++) {
-      const offers = this.offersModel.getByType(this.eventPoints[i].type);
-      const destination = this.destinationsModel.getById(this.eventPoints[i].destination);
-      render(new EventItemView({
-        eventPoint: this.eventPoints[i],
-        offers,
-        destination
+    this.points.forEach((point) => {
+      render(new PointView({
+        point,
+        pointDestination: this.destinationsModel.getById(point.destination),
+        pointOffers: this.offersModel.getByType(point.type)
       }), this.listComponent.getElement());
-    }
+    });
   }
 }
