+++ src/templates/event-list-template.js
@@ -1,0 +1,7 @@
+function createEventListTemplate() {
+  return '<ul class="trip-events__list"></ul>';
+}
+
+export {
+  createEventListTemplate
+};
